==> notebooks/sroie/05_sroie_lsvm.ipynb <==
import os
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.svm import LinearSVC
from sklearn.metrics import classification_report

BASE_PATH = r"C:\Users\Impana\Downloads\invoice-classification\\"
train_path = os.path.join(BASE_PATH, "data", "sroie", "D3_sroie_WNtrain10k.csv")
test_path  = os.path.join(BASE_PATH, "data", "sroie", "D2_sroie_test.csv")

train_df = pd.read_csv(train_path)
test_df  = pd.read_csv(test_path)

print("Train:", train_df.shape, "Test:", test_df.shape)
print("Train cats:", train_df['category'].nunique(), "Test cats:", test_df['category'].nunique())

X_train_text = train_df['text'].astype(str).tolist()
y_train = train_df['category'].astype(str).tolist()

X_test_text = test_df['text_clean'].astype(str).tolist()
y_test = test_df['category'].astype(str).tolist()

vectorizer = TfidfVectorizer(
    ngram_range=(1,2),
    max_features=20000,
    min_df=2
)

X_train = vectorizer.fit_transform(X_train_text)
X_test  = vectorizer.transform(X_test_text)

lsvm = LinearSVC(random_state=42)
lsvm.fit(X_train, y_train)

y_pred = lsvm.predict(X_test)

print(classification_report(y_test, y_pred))
